requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pytest

python-socketio
python-jose
passlib[bcrypt]

uuid
requests
pymongo
bcrypt
PyJWT
pyjwt[crypto]
pandas
numpy
sklearn